refactor: move fetch functionality to it's own folder
